Remove reference to manifest.json
The default template requests a manifest that ends up mapping to a 404
on Mavis, which we assume is what's causing a bug where the Reporting
app pulls in the CSS and JS from Mavis into the frontend.

This also correctly maps the favicons to the ones in the reporting app,
which are the same, but nonetheless it was weird to link to the Mavis
ones.