Allow filtering the reporting dashboard
This wires up the checkboxes and radio buttons in the frontend to the
request that goes to the Totals API in Mavis.

It also adds a button to apply the filters, which will be visible in
situations where JS fails to load.

The year group filter is only partially working at the moment, but it's
because it's not being handled on the Mavis side, which can be done as a
follow-up.

A follow-up PR will add JS to make the form feel better to use, by
submitting it without needing to use the button, and by replacing the
content of the dashboard in-place.