Use htmx to progressively enhance dashboard
This adds `hx-` attributes to the form to trigger a GET request to the
same form endpoint whenever the inputs change. HTMX intercepts this
request, extracts the `#dashboard`, and replaces the `#dashboard` in the
same page.

We also add a `js-hide` CSS class that hides the "Update results" button
if JS is enabled.