Proxy requests to Mavis in development
When running the app locally, the Reporting app lives on
`localhost:4001` and Mavis continues to live on `localhost:4000`.

This is different from production, and means we can't easily tests
features around things like cookie or sessionStorage sharing.

This adds a new `dev_proxy` module that transparently rewrites requests
from `localhost:4001` to `localhost:4000`, making the entire journey
look like it's running on `4001`.