Use ROOT_URL to infer mavis_public_url
Our current setup includes MAVIS_ROOT_URL for internal calls from the
reporting app to Mavis via the API. It's mapped to `web:4000`.

We can't use this for public-facing URLs, like redirects. For that we
need `qa.mavistesting.com` or similar.

We already pass in `qa.mavistesting.com/reports/` as `ROOT_URL`, so we
can remove the `/reports/` suffix and use that instead.