Handle outdated sessions
Logging out of Mavis causes the session cookie in the Reporting app to
become stale. If the user then logs back into Mavis, the Reporting app
will encounter an unhandled 401 as their session is out of date.

This adds a handler to explicitly send the user to Mavis and then back.